:hammer: Move js load requirements out of Updater.
