+++ src/renderer/js/components/Updater.jsx
@@ -1,14 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {map} from 'ramda';
 import {apiRequest} from '../requests';
-import {loadCss, loadScript} from '../../../commons/js';
 import {
     hydrateComponent,
     hydrateProps,
     isComponent,
     prepareProp,
 } from '../hydrator';
+import {loadRequirement, loadRequirements} from '../requirements';
 
 export default class Updater extends React.Component {
     constructor(props) {
@@ -160,7 +159,7 @@ export default class Updater extends React.Component {
                     // reconnect it will do a hard reload of the page api.
                     return this.setState({reloading: true});
                 }
-                this.loadRequirements(filenames, {});
+                filenames.forEach(loadRequirement);
                 break;
             case 'ping':
                 // Just do nothing.
@@ -191,57 +190,6 @@ export default class Updater extends React.Component {
             key: binding.key,
         };
         this.ws.send(JSON.stringify(payload));
-    }
-
-    loadRequirements(requirements, packages) {
-        return new Promise((resolve, reject) => {
-            let loadings = [];
-            // Load packages first.
-            Object.keys(packages).forEach(pack_name => {
-                const pack = packages[pack_name];
-                loadings = loadings.concat(
-                    pack.requirements.map(this.loadRequirement)
-                );
-            });
-            // Then load requirements so they can use packages
-            // and override css.
-            Promise.all(loadings)
-                .then(() => {
-                    let i = 0;
-                    // Load in order.
-                    const handler = () => {
-                        if (i < requirements.length) {
-                            this.loadRequirement(requirements[i]).then(() => {
-                                i++;
-                                handler();
-                            });
-                        } else {
-                            resolve();
-                        }
-                    };
-                    handler();
-                })
-                .catch(reject);
-        });
-    }
-
-    loadRequirement(requirement) {
-        return new Promise((resolve, reject) => {
-            const {url, kind, meta} = requirement;
-            let method;
-            if (kind === 'js') {
-                method = loadScript;
-            } else if (kind === 'css') {
-                method = loadCss;
-            } else if (kind === 'map') {
-                return resolve();
-            } else {
-                return reject({error: `Invalid requirement kind: ${kind}`});
-            }
-            method(url, meta)
-                .then(resolve)
-                .catch(reject);
-        });
     }
 
     _connectWS() {
@@ -298,7 +246,7 @@ export default class Updater extends React.Component {
                 packages: response.packages,
                 requirements: response.requirements,
             });
-            this.loadRequirements(
+            loadRequirements(
                 response.requirements,
                 response.packages
             ).then(() => {
